# docs/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SecureVOTE API",
    "description": "Educational/research prototype of an embedded electronic-voting and election-integrity pipeline. This is NOT production election infrastructure. See /docs for the interactive API explorer.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/elections/{election_id}/audit": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Get Audit Log",
        "description": "Get paginated audit log for an election.",
        "operationId": "get_audit_log_api_elections__election_id__audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/elections/{election_id}/audit/verify": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Verify Audit Chain",
        "description": "Independently verify the audit hash chain.\n\nRecomputes every hash from raw record data \u00e2\u20ac\u201d does NOT read any\nprecomputed verification flag.",
        "operationId": "verify_audit_chain_api_elections__election_id__audit_verify_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/complaints": {
      "get": {
        "tags": [
          "Complaints"
        ],
        "summary": "List Complaints",
        "operationId": "list_complaints_api_complaints_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ComplaintResponse"
                  },
                  "type": "array",
                  "title": "Response List Complaints Api Complaints Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Complaints"
        ],
        "summary": "Submit Complaint",
        "operationId": "submit_complaint_api_complaints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/complaints/{complaint_id}": {
      "get": {
        "tags": [
          "Complaints"
        ],
        "summary": "Get Complaint",
        "operationId": "get_complaint_api_complaints__complaint_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "complaint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Complaint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/complaints/track/{reference_number}": {
      "get": {
        "tags": [
          "Complaints"
        ],
        "summary": "Track Complaint",
        "operationId": "track_complaint_api_complaints_track__reference_number__get",
        "parameters": [
          {
            "name": "reference_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reference Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/complaints/{complaint_id}/status": {
      "patch": {
        "tags": [
          "Complaints"
        ],
        "summary": "Update Complaint Status",
        "operationId": "update_complaint_status_api_complaints__complaint_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "complaint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Complaint Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/elections/{election_id}/polling-stations": {
      "post": {
        "tags": [
          "Polling Stations"
        ],
        "summary": "Create Polling Station",
        "operationId": "create_polling_station_api_elections__election_id__polling_stations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollingStationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollingStationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Polling Stations"
        ],
        "summary": "List Polling Stations",
        "operationId": "list_polling_stations_api_elections__election_id__polling_stations_get",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PollingStationResponse"
                  },
                  "title": "Response List Polling Stations Api Elections  Election Id  Polling Stations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/elections/{election_id}/polling-stations/{station_id}": {
      "get": {
        "tags": [
          "Polling Stations"
        ],
        "summary": "Get Polling Station",
        "operationId": "get_polling_station_api_elections__election_id__polling_stations__station_id__get",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          },
          {
            "name": "station_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollingStationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/elections/{election_id}/polling-stations/{station_id}/status": {
      "patch": {
        "tags": [
          "Polling Stations"
        ],
        "summary": "Update Polling Station Status",
        "operationId": "update_polling_station_status_api_elections__election_id__polling_stations__station_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          },
          {
            "name": "station_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Station Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollingStationStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollingStationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEntryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "election_id": {
            "type": "string",
            "title": "Election Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "event_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Data"
          },
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "sequence_number": {
            "type": "integer",
            "title": "Sequence Number"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "previous_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Hash"
          },
          "entry_hash": {
            "type": "string",
            "title": "Entry Hash"
          }
        },
        "type": "object",
        "required": [
          "id",
          "election_id",
          "event_type",
          "event_data",
          "actor",
          "device_id",
          "sequence_number",
          "timestamp",
          "previous_hash",
          "entry_hash"
        ],
        "title": "AuditEntryResponse"
      },
...
      "AuditLogResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AuditEntryResponse"
            },
            "type": "array",
            "title": "Entries"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "chain_status": {
            "type": "string",
            "title": "Chain Status"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "total",
          "chain_status"
        ],
        "title": "AuditLogResponse"
      },
      "AuditVerificationResponse": {
        "properties": {
          "is_intact": {
            "type": "boolean",
            "title": "Is Intact"
          },
          "total_entries": {
            "type": "integer",
            "title": "Total Entries"
          },
          "verified_entries": {
            "type": "integer",
            "title": "Verified Entries"
          },
          "first_broken_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Broken Index"
          },
          "first_broken_sequence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Broken Sequence"
          },
          "details": {
            "type": "string",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "is_intact",
          "total_entries",
          "verified_entries",
          "details"
        ],
        "title": "AuditVerificationResponse"
      },
...
      "ComplaintCreate": {
        "properties": {
          "election_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Election Id"
          },
          "category": {
            "type": "string",
            "pattern": "^(VOTER_REGISTRATION|POLLING_STATION|EVM|VOTING_ISSUE|ACCESSIBILITY|ELECTION_PROCESS|CANDIDATE_PARTY|TECHNICAL|OTHER)$",
            "title": "Category"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Description"
          },
          "complainant_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Complainant Name"
          },
          "complainant_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complainant Contact"
          }
        },
        "type": "object",
        "required": [
          "category",
          "description",
          "complainant_name"
        ],
        "title": "ComplaintCreate"
      },
      "ComplaintResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "election_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Election Id"
          },
          "reference_number": {
            "type": "string",
            "title": "Reference Number"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "complainant_name": {
            "type": "string",
            "title": "Complainant Name"
          },
          "complainant_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complainant Contact"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "assigned_officer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Officer"
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "election_id",
          "reference_number",
          "category",
          "description",
          "complainant_name",
          "complainant_contact",
          "status",
          "assigned_officer",
          "resolution_notes",
          "created_at",
          "updated_at"
        ],
        "title": "ComplaintResponse"
      },
      "ComplaintStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "pattern": "^(RECEIVED|ASSIGNED|UNDER_REVIEW|RESOLVED|CLOSED)$",
            "title": "Status"
          },
          "assigned_officer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Officer"
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ComplaintStatusUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PollingStationCreate": {
        "properties": {
          "station_code": {
            "type": "string",
            "pattern": "^PS-\\d{3}$",
            "title": "Station Code"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Name"
          },
          "constituency": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Constituency"
          },
          "location": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Location"
          },
          "officer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Officer Name"
          }
        },
        "type": "object",
        "required": [
          "station_code",
          "name",
          "constituency",
          "location"
        ],
        "title": "PollingStationCreate"
      },
      "PollingStationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "election_id": {
            "type": "string",
            "title": "Election Id"
          },
          "station_code": {
            "type": "string",
            "title": "Station Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "constituency": {
            "type": "string",
            "title": "Constituency"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "assigned_devices": {
            "type": "integer",
            "title": "Assigned Devices"
          },
          "registered_voters": {
            "type": "integer",
            "title": "Registered Voters"
          },
          "votes_cast": {
            "type": "integer",
            "title": "Votes Cast"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "officer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Officer Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "election_id",
          "station_code",
          "name",
          "constituency",
          "location",
          "assigned_devices",
          "registered_voters",
          "votes_cast",
          "status",
          "officer_name",
          "created_at"
        ],
        "title": "PollingStationResponse"
      },
      "PollingStationStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "PollingStationStatusUpdate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}